Move pages into subfolder
