+++ src/pages/options/options.html
@@ -31,6 +31,5 @@
             </select>
         </div>
         <!-- <p><a href="docs/help.html" target="_blank">Help</a></p> -->
-        <script src="options.js"></script>
     </body>
 </html>
