+++ src/pages/popup/changelog.html
@@ -6,6 +6,5 @@
     </head>
     <body>
         <div id="push"></div>
-        <script src="changelog.js"></script>
     </body>
 </html>
